Use `ClientPlatform` instead of `String` for analytics entries
